We always want to build against the latest in a mock environment
The builders may have certain packages versionlocked in the host
dnf.conf setup (example: we wanted to pin the builders on a certain
version of ansible-core). This causes trouble in a mock root because it
inherits the builders dnf configuration when trying to fill out a
buildroot.

When running mock, we should always disable the versionlock plugin so
that mockroots get the latest builds from koji.

Fixes: CS-1707

Signed-off-by: Brian Stinson <bstinson@redhat.com>